Fix uint64_t / size_t inconsistency
This might need some further thought. Maybe go uint64_t everywhere
instead.

+++ deltablocks.cpp
@@ -836,7 +836,7 @@ void printTxnFirstConfirmStats(void) {
         tfm::printf("\n");
     }
 }
-Scheduler* deltablocksSetup(uint64_t seed) {
+Scheduler* deltablocksSetup(size_t seed) {
     rng.seed(seed);
     createNodes();
     connectNodes();
